Teams revamp
